Cleanup main.ts
